#234 refactor: 세부 통계 파일 분리
